Models updated, Validator added, Test validation in controller added, Repositories for entities done, logger updated

+++ ComprenteWebApi/src/main/java/com/sachishin/comprente/Repository/model/Admin.java
@@ -1,14 +1,13 @@
 package com.sachishin.comprente.Repository.model;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.sun.istack.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import org.springframework.jmx.export.annotation.ManagedAttribute;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.Table;
+import javax.persistence.*;
 
 @Entity
 @Data
@@ -18,7 +17,7 @@ import javax.persistence.Table;
 public class Admin {
     @Id
     @GeneratedValue
-    private int Id;
+    private long Id;
     @NotNull
     private String email;
     @NotNull
